requirements.txt
streamlit
cryptography
qrcode
pillow
